Fix compiler warnings.
* ospf6d.h (timerstring), watchquagga.c (restart_kill): time_t isn't
  a long everywhere, add casts to long to avoid printf format error.

+++ ospf6d/ospf6d.h
@@ -78,15 +78,15 @@ extern struct thread_master *master;
   do {                                                \
     if ((tv)->tv_sec / 60 / 60 / 24)                  \
       snprintf (buf, size, "%ldd%02ld:%02ld:%02ld",   \
-                (tv)->tv_sec / 60 / 60 / 24,          \
-                (tv)->tv_sec / 60 / 60 % 24,          \
-                (tv)->tv_sec / 60 % 60,               \
-                (tv)->tv_sec % 60);                   \
+                (long)(tv)->tv_sec / 60 / 60 / 24,    \
+                (long)(tv)->tv_sec / 60 / 60 % 24,    \
+                (long)(tv)->tv_sec / 60 % 60,         \
+                (long)(tv)->tv_sec % 60);             \
     else                                              \
       snprintf (buf, size, "%02ld:%02ld:%02ld",       \
-                (tv)->tv_sec / 60 / 60 % 24,          \
-                (tv)->tv_sec / 60 % 60,               \
-                (tv)->tv_sec % 60);                   \
+                (long)(tv)->tv_sec / 60 / 60 % 24,    \
+                (long)(tv)->tv_sec / 60 % 60,         \
+                (long)(tv)->tv_sec % 60);             \
   } while (0)
 #define timerstring_local(tv, buf, size)                  \
   do {                                                    \
